Naively neutralize most overflow: hidden style
The hover effect tooltip is now visible in more situations, like on Dictionary.com.
This still doesn't totally fix the problem, but it makes the effect display correctly in more situations. For example, the hover effect gets partially obscured still if the word takes place in a card, which is common in the search results of Google (ex: search for 'army').